Use consistent structure with editor/save/index
