+++ apps/o2-blocks/src/editor.js
@@ -9,4 +9,4 @@ import './spoiler/editor';
 import './todo/editor';
 import './task';
 import './project-status';
-import './janitorial-state-widget/editor';
+import './janitorial-state-widget';
